Exclude visa-checkout from braintree-android
Visa Checkout uses databinding and seems to be a blocker for the
AndroidX upgrade. Excluding the package from braintree.

This will prevent merchants from processing Visa Checkout until it is
fixed but this is a temporary fix.

Also including a test Card tokenization to verify that Braintree's
tokenization flow works with AndroidX.